requirements.txt
asgiref==3.11.1
cryptography==46.0.5
dj-database-url==3.1.1
Django==6.0.2
django-cors-headers==4.9.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mangum==0.21.0
psycopg==3.3.3
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
whitenoise==6.12.0
twilio
